requirements.txt
apache-airflow==2.4.0
requests
psycopg2
pandas
python-dotenv
